+++ src/relayer/relayer_events.rs
@@ -23,7 +23,7 @@ pub enum RelayerEvents {
     Liquidate {
         liquidation_price: f64,
         available_margin: f64,
-        nonce: u128,
+        nonce: usize,
     },
     CancelOrder {
         uuid: Uuid,
